grainAllocation: move policies to modules (#2518)
Currently, each grain allocation policy is located in the `grainAllocation` module.  This commit seeks to simplify that module by moving policies into their own modules.  This has the added benefit of being able to quickly view each policy.

__Test Plan__
Testing here should be trivial, as existing flow and unit tests should cover any potential errors caused from reorganizing modules.
